Ignore static checks per file.
The idea is that currently the error messages of mypy and pylint are just overwhelming so they're probably ignored by everyone (I certainly ignore them). Hence I think it better to turn the process around: currently, per default, checks are disabled for all files. When you modify a file extensively and feel it's now in a good state turn on static checking to keep it that way.