Revert "Run tests on all branches."
This leads to the tests always running twice (on the branch, rebased onto master) for pull requests. We don't want that. To address the issues of not automatically having tests run on branches the answer is create a pull request (leave it in a draft state if it's still work in progress).