Move `massFlowSolver` to `trnsysGUI`.
All the code should be under one top-level directory package. `trnsysGUI` might not be the best name for it, but we can always rename it in the future.

Incidentally, because `massFlowSolver` was outside `trnsysGUI` before it wasn't type checked. It automatically is now. That's also why I had to make a few seemingly unrelated changes to satisfy `mypy` and friends.