Don't create as build and upload it as artifact to pull requests.
...and create builds on `Etna`.

Artifacts for pull requests are currently broken as they are always on against `master` (due to restrictions of `https://github.com/peter-evans/repository-dispatch`/`repository_dispatch` in general),
and not on the branch of the pull request.

`repository_dispatch` doesn't seem to be the right tool for this, I think we should look into callable workflows going forward.